More dry-run mode changes on weekly.js

+++ weekly.js
@@ -321,21 +321,31 @@ function createWeeklyContentFromYaml({ configYaml, filename }) {
     schedule,
     title 
   } = yaml.parse(configYaml);
-  const weeklyFolder       = path.join("curriculum", filename);
-  const weeklyFolderExists = fs.existsSync(weeklyFolder);
+  const weeklyFolder           = path.join("curriculum", filename);
+  const weeklyFolderExists     = fs.existsSync(weeklyFolder);
   const weeklyUserFolder       = path.join("user", filename);;
   const weeklyUserFolderExists = fs.existsSync(weeklyUserFolder);
+  const isDryRunMode           = global.sgenConfig.dryRun;
 
   if ( weeklyFolderExists ) {
 
     warn(`Folder '${weeklyFolder}' already exists. (Selectively) deleting it to begin from scratch.`);
-    // fs.rmSync(weeklyFolder, { recursive: true });
-    rmSyncExclude(weeklyFolder, ["WEEKEND.md"]); // Selective rm excluding files/folders inside the 2nd argument
+    
+    if ( isDryRunMode ){
+      console.log(`[DRY-RUN MODE] Deleting the contents of folder '${weeklyFolder}' excluding the following: WEEKEND.md`);
+    } else {
+      // fs.rmSync(weeklyFolder, { recursive: true });
+      rmSyncExclude(weeklyFolder, ["WEEKEND.md"]); // Selective rm excluding files/folders inside the 2nd argument
+    }
     
   } else {
 
-    fs.mkdirSync(weeklyFolder);
-    info(`Folder '${weeklyFolder}' created.`);
+    if ( isDryRunMode ){
+      console.log(`[DRY-RUN MODE] Folder '${weeklyFolder}' created.`);
+    } else {
+      fs.mkdirSync(weeklyFolder);
+      info(`Folder '${weeklyFolder}' created.`);
+    }
   
   }
